% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{anti_pd1_favor_signature}
\alias{anti_pd1_favor_signature}
\title{Runs an anti-PD1 favor signature analysis}
\usage{
anti_pd1_favor_signature(z_score_matrix_file)
}
\arguments{
\item{z_score_matrix_file}{A matrix of z-scores for all genes in all tumor samples and regions. Can be created using function z_score_calculations.}
}
\description{
Takes in a z-score matrix created by function z_score_calculations
Outputs a table of anti-PD1 favor scores for each tumor sample or region and a heatmap of the anti-PD1 favor scores
}
