% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaITH.R
\name{apoptosis_signature}
\alias{apoptosis_signature}
\title{Runs an apoptosis signature analysis}
\usage{
apoptosis_signature(z_score_matrix_file)
}
\arguments{
\item{z_score_matrix_file}{A matrix of z-scores for all genes in all tumor samples and regions. Can be created using function z_score_calculations.}
}
\description{
Takes in a z-score matrix created by function z_score_calculations
Outputs a table of apoptosis scores for each tumor sample or region and a heatmap of the apoptosis scores
}
