% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{dna_dendrograms}
\alias{dna_dendrograms}
\title{Creates DNA dendrogram}
\usage{
dna_dendrograms(dna_dendro_list_file)
}
\arguments{
\item{dna_dendro_list_file}{A text file containing a list of sample files, one file name per line. Each file contains a matrix with variant allele frequency of all the variations in each tumor region and normal.}
}
\description{
Takes in a text file containing list of sample files. Each file contains a matrix with variant allele frequency of all the variations in each tumor region and normal.
Creates a DNA distance matrix, containing the distance between all the tumor regions and normal, a DNA tree topology file, and an unrooted DNA dendrogram.
}
