% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaITH.R
\name{dna_phylograms}
\alias{dna_phylograms}
\title{Creates DNA phylogram}
\usage{
dna_phylograms(dna_phylo_list_file)
}
\arguments{
\item{dna_phylo_list_file}{A text file containing a list of sample files, one file name per line. Each file contains a matrix with variant allele frequency of all the variations in each tumor region and normal.}
}
\description{
Takes in a text file containing list of sample files. Each file contains a matrix with variant allele frequency of all the variations in each tumor region and normal.
Creates a DNA distance matrix, containing the distance between all the tumor regions and normal, a DNA tree topology file, and an unrooted DNA phylogram.
}
