% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaITH.R
\name{emt_signature}
\alias{emt_signature}
\title{Runs an emt signature analysis}
\usage{
emt_signature(z_score_matrix_file)
}
\arguments{
\item{z_score_matrix_file}{A matrix of z-scores for all genes in all tumor samples and regions. Can be created using function z_score_calculations.}
}
\description{
Takes in a z-score matrix created by function z_score_calculations
Outputs a table of emt(epithelial-mesenchymal transition) scores for each tumor sample or region and a heatmap of the emt scores
}
