% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{immune_dendrograms}
\alias{immune_dendrograms}
\title{Creates Immune dendrogram}
\usage{
immune_dendrograms(immune_dendro_list_file)
}
\arguments{
\item{immune_dendro_list_file}{A text file containing a list of sample files, one file name per line. Each file contains a matrix with proportion of immune cells (inferred by CIBERSORT) of all immune cell types in each tumor region and normal.}
}
\description{
Takes in a text file containing list of sample files. 
Creates an immune distance matrix, containing the distance between all the tumor regions and normal, an immune tree topology file, and an unrooted phylogram.
}
