% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaITH.R
\name{multi_level_divergence_diversity}
\alias{multi_level_divergence_diversity}
\title{Performs a multi-level divergence and diversity analysis}
\usage{
multi_level_divergence_diversity(sample_names_file)
}
\arguments{
\item{sample_names_file}{A text file containing list of sample names, one file name per line}
}
\description{
Takes in text file containing list of samples, one sample per line, and distance matrix files generated by dna_phylograms, rna_phylograms, and immune_phylograms methods.
Outputs a png file containing 6 plots, 3 diversity/divergence plots and 3 diversity plots.
NOTE: Requires output of functions dna_phylograms, rna_phylograms, and immune_phylograms to be in same folder
}
