% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{rna_dendrograms}
\alias{rna_dendrograms}
\title{Creates RNA dendrogram}
\usage{
rna_dendrograms(rna_phylo_list_file)
}
\arguments{
\item{rna_dendro_list_file}{A text file containing a list of sample files, one file name per line. Each file contains a matrix with expression values (preferably log2(TPM+1)) of all the genes in each tumor region and normal.}
}
\description{
Takes in a text file containing list of sample files. 
Creates a RNA distance matrix, containing the distance between all the tumor regions and normal, a RNA tree topology file, and a RNA unrooted dendrogram.
}
