% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_analysis_function.R
\name{signature_analysis}
\alias{signature_analysis}
\title{Runs a signature analysis on an RNA expression matrix}
\usage{
signature_analysis(rna_expression_matrix_file)
}
\arguments{
\item{rna_expression_matrix_file}{File containing expression level of all genes in all tumor samples. Header of first column should be "Gene" and each normal file should end in ".nr"}
}
\description{
Takes in RNA expression matrix containing the expression level of all genes in all tumor samples and normals.
Creates a z-score matrix and uses this matrix to find the hypoxia, proliferation, apoptosis, drug-resistance, emt, and anti-PD1 favor scores for each tumor sample or region. Outputs a table and heatmap of these scores.
}
