% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaITH.R
\name{snv_heatmaps}
\alias{snv_heatmaps}
\title{Outputs a heatmap of variant allele frequency.}
\usage{
snv_heatmaps(dna_phylo_list_file)
}
\arguments{
\item{dna_phylo_list_file}{A text file containing a list of sample files, one file name per line. Each file contains a matrix with variant allele frequency of all the variations in each tumor region and normal.}
}
\description{
Takes in text file containing list of sample files.  
Outputs a heatmap of variant allele frequency (vaf) of all variations for each sample.
}
