% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{specified_geneset_signature}
\alias{specified_geneset_signature}
\title{Runs a signature analysis using a user-supplied geneset}
\usage{
specified_geneset_signature(z_score_matrix_file, geneset)
}
\arguments{
\item{z_score_matrix_file}{A matrix of z-scores for all genes in all tumor samples and regions. Can be created using function z_score_calculations.}

\item{geneset}{A text file containing a list of genes supplied by the user, one gene per line.}
}
\description{
Takes in a z-score matrix created by function z_score_calculations and a geneset
Outputs a table of geneset scores and a geneset scores heatmap
}
