% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaITH.R
\name{z_score_calculations}
\alias{z_score_calculations}
\title{Generates a matrix containing z-scores}
\usage{
z_score_calculations(rna_expression_matrix_file)
}
\arguments{
\item{rna_expression_matrix_file}{File containing expression level of all genes in all tumor samples. Header of first column should be "Gene" and each normal file should end in ".nr"}
}
\description{
Takes in RNA expression matrix containing the expression level of all genes in all tumor samples and normals.
Creates a matrix(tab-delimited text file) containing z-scores for all genes in all of the tumor samples or regions and normals.
}
