% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaITH_analysis_function.R
\name{metaITH_analysis}
\alias{metaITH_analysis}
\title{Runs a comprehensive intratumor heterogeneity analysis}
\usage{
metaITH_analysis(dna_dendro_list, rna_dendro_list, immune_dendro_list, multi_region_sample_names, rna_expression_matrix)
}
\arguments{
\item{dna_dendro_list}{A text file containing a list of sample files, one file name per line. Each file contains a matrix with variant allele frequency of all the variations in each tumor region and normal.}

\item{rna_dendro_list}{A text file containing a list of sample files, one file name per line. Each file contains a matrix with expression values (preferably log2(TPM+1)) of all the genes in each tumor region and normal.}

\item{immune_dendro_list}{A text file containing a list of sample files, one file name per line. Each file contains a matrix with proportion of immune cells (inferred by CIBERSORT) of all immune cell types in each tumor region and normal.}

\item{multi_region_sample_names}{A text file containing list of sample names for multi-region diversity and divergence analysis, one file name per line}

\item{rna_expression_matrix}{File containing expression level of all genes in all tumor samples. Header of first column should be "Gene" and each normal file should end in ".nr"}
}
\description{
Analyzes several sample files by running a dendrogram analysis (using DNA, RNA, and immune expression data), signature analysis (using RNA expression data), and multi-region diversity and divergence analysis (using DNA, RNA, and immune expression data)
}
